requirements.txt
astronomer-cosmos==1.10.0
boto3
psycopg2-binary
pydub==0.25.1
google-genai
